webhook fixes (#1609)
* webhook fixes

* billing updates

* using make request

* using float for integer

* cleanups

+++ lib/glific_web/schema/billing_types.ex
@@ -36,7 +36,7 @@ defmodule GlificWeb.Schema.BillingTypes do
     field :currency, :string
     field :is_active, :boolean
     field :deduct_tds, :boolean
-    field :tds_amount, :integer
+    field :tds_amount, :float
     field :stripe_customer_id, :string
     field :stripe_payment_method_id, :string
     field :stripe_subscription_id, :string
@@ -53,7 +53,7 @@ defmodule GlificWeb.Schema.BillingTypes do
     field :email, :string
     field :currency, :string
     field :deduct_tds, :boolean
-    field :tds_amount, :integer
+    field :tds_amount, :float
     field :stripe_subscription_status, :string
     field :stripe_subscription_id, :string
   end
